+++ sites/geohub/src/routes/api/datasets/[id]/+server.ts
@@ -2,10 +2,10 @@ import type { RequestHandler } from './$types'
 import pkg from 'pg'
 const { Pool } = pkg
 
-import { DATABASE_CONNECTION } from '$lib/server/variables/private'
+import { env } from '$env/dynamic/private'
 import { generateAzureBlobSasToken } from '$lib/server/helpers'
 import type { DatasetFeature, Tag } from '$lib/types'
-const connectionString = DATABASE_CONNECTION
+const connectionString = env.DATABASE_CONNECTION
 
 export const GET: RequestHandler = async ({ params, locals }) => {
   const session = await locals.getSession()