+++ sites/geohub/src/routes/api/stac/mosaicjson/+server.ts
@@ -1,12 +1,12 @@
 import type { StacCollection, StacItemFeatureCollection } from '$lib/types'
 import type { RequestHandler } from './$types'
-import { PUBLIC_TITILER_ENDPOINT } from '$lib/variables/public'
+import { PUBLIC_TITILER_ENDPOINT } from '$env/static/public'
 
 const TITILER_MOSAIC_ENDPOINT = PUBLIC_TITILER_ENDPOINT.replace('cog', 'mosaicjson')
 
 import { BlobServiceClient, StorageSharedKeyCredential } from '@azure/storage-blob'
-import { AZURE_STORAGE_ACCOUNT, AZURE_STORAGE_ACCESS_KEY } from '$lib/server/variables/private'
-const sharedKeyCredential = new StorageSharedKeyCredential(AZURE_STORAGE_ACCOUNT, AZURE_STORAGE_ACCESS_KEY)
+import { env } from '$env/dynamic/private'
+const sharedKeyCredential = new StorageSharedKeyCredential(env.AZURE_STORAGE_ACCOUNT, env.AZURE_STORAGE_ACCESS_KEY)
 
 import fs from 'fs'
 import path from 'path'