+++ sites/geohub/src/routes/api/vector/[source]/metadata.json/+server.ts
@@ -7,7 +7,7 @@ import {
   generateMetadataJson,
   getPgtileservTileJson,
 } from '$lib/server/helpers'
-import { PUBLIC_MARTIN_API_ENDPOINT, PUBLIC_PGTILESERV_API_ENDPOINT } from '$lib/variables/public'
+import { env } from '$env/dynamic/private'
 
 /**
  * /api/vector/[source]/metadata.json?table={table}&type={type}&pbfpath={pbfpath}
@@ -33,7 +33,7 @@ export const GET: RequestHandler = async ({ url, params }) => {
       metadatajson = await getStaticPbfMetadataJson(url.origin, pbfpath)
       break
     case 'martin':
-      tilejson = await getMartinTileJson(table, PUBLIC_MARTIN_API_ENDPOINT)
+      tilejson = await getMartinTileJson(table, env.MARTIN_API_ENDPOINT)
       if (!tilejson) {
         return new Response(JSON.stringify({ message: `table: ${table} not found.` }), {
           status: 404,
@@ -42,7 +42,7 @@ export const GET: RequestHandler = async ({ url, params }) => {
       metadatajson = await generateMetadataJson(tilejson)
       break
     case 'pgtileserv':
-      tilejson = await getPgtileservTileJson(table, type, PUBLIC_PGTILESERV_API_ENDPOINT)
+      tilejson = await getPgtileservTileJson(table, type, env.PGTILESERV_API_ENDPOINT)
       metadatajson = await generateMetadataJson(tilejson)
       break
     default: